Article about BMS LV
In case of this particular BMS, there is a self-made battery pack with 21700 cells in 6S4P configuration (max 25,2V) that will be managed. It should be noted that many design decisions are dictated by the regulations that we have to follow when building a racing car. Low voltage battery powers all the systems in the car.
Some important functions this PCB is incuding:
• Voltage monitoring on individual cells (6S – 6 measurement points),
• Temperature monitoring on 33% of all cells,
• Battery current consumption monitoring,
• Battery protection against critical errors specified in the regulations of Formula Student Germany,
• Battery cell balancing system while charging,
• Information processing and sending via serial port and CAN,
• Voltage measurement point behind a fuse

PCB manufactring supported by https://jlcpcb.com/HAR

Author: Maksymilian Jaruga
Schematics and more:
https://github.com/PUT-Motorsport/PUTM_EV_BM